tiny refactoring and added admin entities. some bugs coming up
