+++ entities/src/mod.rs
@@ -13,5 +13,6 @@ pub mod rpc_key;
 pub mod rpc_request;
 pub mod sea_orm_active_enums;
 pub mod secondary_user;
+pub mod serialization;
 pub mod user;
 pub mod user_tier;
